--- 3.AspNetCoreWebApi/Metrics/MetricsAgent/DAL/Requests/RamMetricCreateRequest.cs
@@ -1,18 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace MetricsAgent.DAL.Requests
-{
-    public class RamMetricCreateRequest
-    {
-        public int Value { get; set; }
-        public DateTimeOffset Time { get; set; }
-
-        public override string ToString()
-        {
-            return $"{{Value={Value}, Time={Time}}}";
-        }
-    }
-}
