+++ 3.AspNetCoreWebApi/Metrics/MetricsAgent/Requests/CpuMetricCreateRequest.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.DAL.Requests
+namespace MetricsAgent.Requests
 {
     public class CpuMetricCreateRequest
     {
