+++ 3.AspNetCoreWebApi/Metrics/MetricsAgent/Requests/NetworkMetricCreateRequest.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.DAL.Requests
+namespace MetricsAgent.Requests
 {
-    public class HddMetricCreateRequest
+    public class NetworkMetricCreateRequest
     {
         public int Value { get; set; }
         public DateTimeOffset Time { get; set; }
