+++ 3.AspNetCoreWebApi/Metrics/MetricsAgent/Responses/DotNetMetricResponse.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.DAL.Responses
+namespace MetricsAgent.Responses
 {
     public class DotNetMetricResponse
     {
-        public List<DotNetMetricDTO> Metrics { get; set; } = new List<DotNetMetricDTO>();
+        public List<DotNetMetricDto> Metrics { get; set; } = new List<DotNetMetricDto>();
     }
 }
