+++ 3.AspNetCoreWebApi/Metrics/MetricsAgent/Responses/HddMetricDTO.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.DAL.Responses
+namespace MetricsAgent.Responses
 {
-    public class DotNetMetricDTO
+    public class HddMetricDto
     {
         public int Id { get; set; }
