+++ 3.AspNetCoreWebApi/Metrics/MetricsAgent/Responses/HddMetricResponse.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.DAL.Responses
+namespace MetricsAgent.Responses
 {
     public class HddMetricResponse
     {
-        public List<HddMetricDTO> Metrics { get; set; } = new List<HddMetricDTO>();
+        public List<HddMetricDto> Metrics { get; set; } = new List<HddMetricDto>();
     }
 }
