+++ 3.AspNetCoreWebApi/Metrics/MetricsAgent/Responses/NetworkMetricDTO.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace MetricsAgent.Responses
+{
+    public class NetworkMetricDto
+    {
+        public int Id { get; set; }
+
+        public int Value { get; set; }
+
+        public DateTimeOffset Time { get; set; }
+    }
+}
