+++ 3.AspNetCoreWebApi/Metrics/MetricsAgent/Responses/NetworkMetricResponse.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.DAL.Responses
+namespace MetricsAgent.Responses
 {
     public class NetworkMetricResponse
     {
-        public List<NetworkMetricDTO> Metrics { get; set; } = new List<NetworkMetricDTO>();
+        public List<NetworkMetricDto> Metrics { get; set; } = new List<NetworkMetricDto>();
     }
 }
