+++ 3.AspNetCoreWebApi/Metrics/MetricsAgent/Responses/RamMetricResponse.cs
@@ -3,10 +3,10 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace MetricsAgent.DAL.Responses
+namespace MetricsAgent.Responses
 {
     public class RamMetricResponse
     {
-        public List<RamMetricDTO> Metrics { get; set; } = new List<RamMetricDTO>();
+        public List<RamMetricDto> Metrics { get; set; } = new List<RamMetricDto>();
     }
 }
